Changed the default shell. Added shell specific section to support zsh.
The default shell is now /bin/sh instead of zsh. For zsh, we use a
switch to set the flag so arrays are handled correctly.

Technically, we could use the same technique to do similar things with
other shells. This was only required for zsh at this point, but the
offer stands.